emotion_detection/requirements.txt
huggingface_hub
streamlit
python-dotenv